sql: move index selection code to new files
